requirements.txt
openai>=1.30.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
textstat>=0.7.3
pandas>=2.2.0
rich>=13.7.0
tqdm>=4.66.0
pydantic>=2.7.0
